Actualizado con todos los cambios de la API de pruebas 08/01/2025

+++ src/main/java/dam/backend/controller/ListaController.java
@@ -7,8 +7,9 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
+//import org.springframework.security.core.Authentication;
+//import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import dam.backend.domain.Lista;
 import dam.backend.domain.Usuario;
@@ -25,6 +27,7 @@ import dam.backend.repository.ListaRepository;
 import dam.backend.repository.UsuarioRepository;
 
 @RestController
+@CrossOrigin(origins = "http://localhost:8100, http://localhost:8101")
 @RequestMapping("/api/lista")
 public class ListaController {
 
@@ -32,19 +35,27 @@ public class ListaController {
 	ListaRepository listaRepository;
 	@Autowired
 	UsuarioRepository usuarioRepository;
-	
 	/*
 	@GetMapping({"/",""}) 
 	public List <Lista> index() {
 		return listaRepository.findAll();
-	}*/
+	}
+	*/
+	//Devuelve pelicula por ID
+	@GetMapping("/{id}")
+	public Lista show(@PathVariable("id") int id) { 
+		return listaRepository.findById(id).orElse(null);
+	}
 	
+	
+	//Metodo para devolver las listas de un usuario
+	//Devuelve una lista de listas
 	@GetMapping("")
 	public List <Lista> index(Principal principal) {
-		System.out.println(principal.getName());
+		/*System.out.println(principal.getName());
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 		String currentPrincipalName = authentication.getName();
-		System.out.println(currentPrincipalName);
+		System.out.println(currentPrincipalName);*/
 		Optional<Usuario> usuario = usuarioRepository.findByEmail(principal.getName());
 		
 		
@@ -56,35 +67,59 @@ public class ListaController {
 		}
 		return new ArrayList();
 	}
-	//Devuelve pelicula por ID
-	@GetMapping("/id/{id}")
-	public Lista show(@PathVariable int id) { 
-		return listaRepository.findById(id).orElse(null);
-	}
 	
-	//Crear pelicula
-	@PostMapping({"crea"})
+	
+	/*****
+	 * Crea una lista
+	 * Le pasamos el parametro Principal que trae la informacion del usuario
+	 */
+	@PostMapping("")
 	@ResponseStatus (HttpStatus.CREATED)
-	public Lista creaLista(@RequestBody Lista lista) {  
-		return listaRepository.save(lista);
+	public List<Lista> creaLista(@RequestBody Lista lista, Principal principal){
+		Optional<Usuario> usuario = usuarioRepository.findByEmail(principal.getName());
+		
+		if(usuario.isPresent()) {
+			listaRepository.save(lista);
+			
+			Optional<List <Lista>> listas = listaRepository.findByUsuario(usuario.get());
+			
+			if(listas.isPresent()) {
+				return listas.get();
+			}
+		}
+		
+		return new ArrayList<Lista>();
 	}
 	
 	/***
 	 * Borramos una lista
-	 * Le pasamos por parametro el ID
+	 * Le pasamos por parametro el ID y un Principal con informacion del usuario
 	 * **/
-	@DeleteMapping("borra/{id}")
+	@DeleteMapping("/{id}")
 	@ResponseStatus (HttpStatus.NO_CONTENT)
-	public void borraLista(@PathVariable int id) {
-		listaRepository.deleteById(id);
+	public List<Lista> borraLista(@PathVariable("id") int id, Principal principal) {
+		Optional<Usuario> usuario = usuarioRepository.findByEmail(principal.getName());
+		
+		if (usuario.isPresent()){
+			listaRepository.deleteById(id);
+			
+			Optional<List <Lista>> listas = listaRepository.findByUsuario(usuario.get());
+			
+			if(listas.isPresent()) {
+				return listas.get();
+			}else {
+	            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Lista no encontrada o no pertenece al usuario");
+	        }
+		}
+		return new ArrayList<Lista>();
 	}
 	
 	/**
 	 * Actualizamos una lista dependiendo de su id
 	 * **/
-	@PutMapping("actualiza/{id}")
+	@PutMapping("/{id}")
 	@ResponseStatus (HttpStatus.CREATED)
-	public Lista actualizaLista(@RequestBody Lista lista, @PathVariable int id) {
+	public Lista actualizaLista(@RequestBody Lista lista, @PathVariable("id") int id) {
 		Lista actuLista = listaRepository.findById(id).orElse(null);
 		
 		actuLista.setUsuario(lista.getUsuario());
